Refactor addCouponCode
